Some minor improvements to axis and stacked area. - Allow tic step to be a float. - Make StackedArea.SetData take a float[][]. This eliminated a warning about conversions. - Make StackedArea an IPrimerGraphData so its Transition method will be called with Graph3.Transition.

+++ Primer.Graph/Axis3/Axis3.cs
@@ -234,7 +234,7 @@ namespace Primer.Graph
 
         #region Tics
 
-        public int ticStep = 2;
+        public float ticStep = 2;
         public bool showZero;
         public int labelNumberOffset;
         public AxisTic ticPrefab;
